python_projects/029_project/requirements.txt
opencv-python>=4.5.0
pyzbar>=0.1.8
python-barcode>=0.13.1
pandas>=1.3.0
openpyxl>=3.0.7
matplotlib>=3.5.0
Pillow>=8.0.0 